extracts bullets business logic to a hook
